+++ pom.xml
@@ -9,7 +9,7 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <maven.compiler.source>20</maven.compiler.source>
         <maven.compiler.target>20</maven.compiler.target>
-        <exec.mainClass>mimikko.zazalng.puddle.Puddle</exec.mainClass>
+        <exec.mainClass>mimikko.zazalng.pudel.Puddle</exec.mainClass>
     </properties>
     <build>
         <plugins>
@@ -21,7 +21,7 @@
                     <archive>
                         <manifest>
                             <addClasspath>true</addClasspath>
-                            <mainClass>mimikko.zazalng.puddle.Puddle</mainClass>
+                            <mainClass>mimikko.zazalng.pudel.Puddle</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
