--- src/main/java/mimikko/zazalng/puddle/commands/Command.java
@@ -1,10 +1,0 @@
-package mimikko.zazalng.puddle.commands;
-
-import mimikko.zazalng.puddle.entities.GuildEntity;
-import mimikko.zazalng.puddle.entities.UserEntity;
-
-public interface Command {
-    void execute(GuildEntity guild, UserEntity user, String replyChannel, String args);
-    String getDescription();
-    String getDetailedHelp();
-}
