--- src/main/java/mimikko/zazalng/puddle/commands/music/MusicPlaying.java
@@ -1,23 +1,0 @@
-package mimikko.zazalng.puddle.commands.music;
-
-import mimikko.zazalng.puddle.commands.Command;
-import mimikko.zazalng.puddle.entities.GuildEntity;
-import mimikko.zazalng.puddle.entities.UserEntity;
-
-public class MusicPlaying implements Command {
-
-    @Override
-    public void execute(GuildEntity guild, UserEntity user, String replyChannel, String args) {
-
-    }
-
-    @Override
-    public String getDescription() {
-        return "";
-    }
-
-    @Override
-    public String getDetailedHelp() {
-        return "";
-    }
-}
