--- src/main/java/mimikko/zazalng/puddle/utility/IntegerUtility.java
@@ -1,11 +1,0 @@
-package mimikko.zazalng.puddle.utility;
-
-public class IntegerUtility {
-    public static int randomInt(int max){
-        return (int) (Math.random()*(max));
-    }
-
-    public static int randomInt(int min, int max){
-        return (int) (Math.random()*(max - min) + min);
-    }
-}
