--- src/main/java/mimikko/zazalng/puddle/utility/StringUtility.java
@@ -1,7 +1,0 @@
-package mimikko.zazalng.puddle.utility;
-
-public class StringUtility {
-    public static boolean isEqualStr(String str, String arg){
-        return str.equals(arg);
-    }
-}
