+++ src/main/java/mimikko/zazalng/pudel/commands/Command.java
@@ -1,0 +1,10 @@
+package mimikko.zazalng.pudel.commands;
+
+import mimikko.zazalng.pudel.entities.GuildEntity;
+import mimikko.zazalng.pudel.entities.UserEntity;
+
+public interface Command {
+    void execute(GuildEntity guild, UserEntity user, String replyChannel, String args);
+    String getDescription();
+    String getDetailedHelp();
+}
