+++ src/main/java/mimikko/zazalng/pudel/handlers/AudioTrackHandler.java
@@ -1,0 +1,62 @@
+package mimikko.zazalng.pudel.handlers;
+
+import com.sedmelluq.discord.lavaplayer.player.AudioPlayer;
+import com.sedmelluq.discord.lavaplayer.player.event.AudioEventAdapter;
+import com.sedmelluq.discord.lavaplayer.tools.FriendlyException;
+import com.sedmelluq.discord.lavaplayer.track.AudioTrack;
+import com.sedmelluq.discord.lavaplayer.track.AudioTrackEndReason;
+import mimikko.zazalng.pudel.manager.MusicManager;
+
+public class AudioTrackHandler extends AudioEventAdapter {
+    private final MusicManager musicManager;
+
+    public AudioTrackHandler(MusicManager musicManager) {
+        this.musicManager = musicManager;
+    }
+
+    public MusicManager getMusicManager(){
+        return this.musicManager;
+    }
+
+    @Override
+    public void onPlayerPause(AudioPlayer player) {
+        // Player was paused
+    }
+
+    @Override
+    public void onPlayerResume(AudioPlayer player) {
+        // Player was resumed
+    }
+
+    @Override
+    public void onTrackStart(AudioPlayer player, AudioTrack track) {
+        // A track started playing
+    }
+
+    @Override
+    public void onTrackEnd(AudioPlayer player, AudioTrack track, AudioTrackEndReason endReason) {
+        if (endReason.mayStartNext) {
+            // Start next track
+            getMusicManager().nextTrack(false);
+        } else if(endReason == AudioTrackEndReason.STOPPED){
+
+        }
+
+        // endReason == FINISHED: A track finished or died by an exception (mayStartNext = true).
+        // endReason == LOAD_FAILED: Loading of a track failed (mayStartNext = true).
+        // endReason == STOPPED: The player was stopped.
+        // endReason == REPLACED: Another track started playing while this had not finished
+        // endReason == CLEANUP: Player hasn't been queried for a while, if you want you can put a
+        //                       clone of this back to your queue
+    }
+
+    @Override
+    public void onTrackException(AudioPlayer player, AudioTrack track, FriendlyException exception) {
+        // An already playing track threw an exception (track end event will still be received separately)
+    }
+
+    @Override
+    public void onTrackStuck(AudioPlayer player, AudioTrack track, long thresholdMs) {
+        // Audio track has been unable to provide us any audio, might want to just start a new track
+    }
+}
