+++ src/main/java/mimikko/zazalng/pudel/utility/IntegerUtility.java
@@ -1,0 +1,11 @@
+package mimikko.zazalng.pudel.utility;
+
+public class IntegerUtility {
+    public static int randomInt(int max){
+        return (int) (Math.random()*(max));
+    }
+
+    public static int randomInt(int min, int max){
+        return (int) (Math.random()*(max - min) + min);
+    }
+}
