+++ src/main/java/mimikko/zazalng/pudel/utility/StringUtility.java
@@ -1,0 +1,7 @@
+package mimikko.zazalng.pudel.utility;
+
+public class StringUtility {
+    public static boolean isEqualStr(String str, String arg){
+        return str.equals(arg);
+    }
+}
